refactor(log): switch log dma to retarget
