+++ app/model_car/components/CMakeLists.txt
@@ -3,11 +3,8 @@
 target_sources(${CMAKE_PROJECT_NAME} PRIVATE
     # Add user sources here
     ../../../components/ring_buf/ring_buf.c
-    # ../../../components/log/log.c
+    ../../../components/log/log.c
     ../../../components/shell/shell.c
-
-    log/log.c
-    log/log_dma.c
 
     shell/shell.c
     shell/shell_sys.c
